requirements.txt
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
importlib-metadata==3.8.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
passlib==1.7.4
pathspec==0.8.1
protobuf==3.15.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
PyMySQL==1.0.2
python-jose==3.2.0
python-multipart==0.0.5
regex==2021.3.17
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.3
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
zipp==3.4.1
